The documentation did not compilecd because of the way in which things were commented out. What we had in root.lexc was:
!+FutDf !!= * @CODE@ Indicates Future Definite tense
!+FutIt !!= * @CODE@ Indicates Future Intentional tense
!+FutCn !!= * @CODE@ Indicates Future Conditional

After compiling the root-morphology.jspwiki, the result was:

 * !+FutDf  Indicates Future Definite tense
 * !+FutIt  Indicates Future Intentional tense
 * !+FutCn  Indicates Future Conditional

(the "!+FutDf! part was placed to the left of !!=, where "@CODE@" is).

This results in an illegal jspwiki text: exclamation mark
is level-three headline, whereas star is bullet list, and 
a headline cannot at the same time be a bullet list.

So, if we want to comment out definitions, we must remove the !!=.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/crk@93557 c7155fb1-f0a7-4240-a2fc-2600b6f42f90