Verb status finally:
Total passes: 1124, Total fails: 0, Total: 1124

What had happened was that we had this regex:

< "+Fut+Int":0 "@U.verb.sbjFutureInt@" > 1sgsubjunctive ;  !"

i.e., with +Fut+Ind as __one__ Multichar_Symbol.
Thus, we was not able to generate anything, since our
input was not two symbols, but one.

Now, they are kept separate:

< "+Fut":0 "+Int":0 "@U.verb.sbjFutureInt@" > 1sgsubjunctive ;  !"

i.e., +Fut and +Int are atomic symbols, and generation works.